requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blockbuster==1.5.24
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cloudpickle==3.1.1
cryptography==44.0.3
distro==1.9.0
fastapi==0.115.12
forbiddenfruit==0.1.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
langchain==0.3.25
langchain-core==0.3.60
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langgraph==0.4.5
langgraph-api==0.2.29
langgraph-checkpoint==2.0.26
langgraph-cli==0.2.10
langgraph-prebuilt==0.1.8
langgraph-runtime-inmem==0.1.0
langgraph-sdk==0.1.69
langsmith==0.3.42
openai==1.79.0
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.1.3
starlette==0.46.2
structlog==25.3.0
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
truststore==0.10.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
xxhash==3.5.0
zstandard==0.23.0
